servantx-backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
python-multipart==0.0.12
sendgrid==6.11.0
python-dotenv==1.0.0
email-validator==2.1.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
requests==2.31.0
pydantic[email]==2.12.5
python-jose[cryptography]==3.3.0
PyJWT==2.9.0
sqlalchemy>=2.0.36
asyncpg>=0.29.0
psycopg>=3.1.18
psycopg2-binary>=2.9.9
greenlet>=3.0.3
PyPDF2==3.0.1
openai>=1.66.0
pydantic-settings==2.2.1
slack-sdk==3.23.0
langfuse==2.0.1
jinja2==3.1.2
alembic>=1.14.0
celery==5.4.0
redis==5.0.8
pytest==8.3.4
python-docx==1.1.2

